+++ spec/aws/active_job/sqs/poller_spec.rb
@@ -13,48 +13,42 @@ module Aws
 
         before do
           allow(ActiveSupport::Logger).to receive(:new).and_return(logger)
-          Aws::ActiveJob::SQS.config.client = sqs_client
+          allow_any_instance_of(Configuration).to receive(:client).and_return(sqs_client)
         end
 
         describe '#initialize' do
-          it 'parses args' do
-            poller = Poller.new(['--queue', 'default', '-v', '360'])
+          it 'initializes options' do
+            poller = Poller.new(max_messages: 3, visibility_timeout: 360)
             parsed = poller.instance_variable_get(:@options)
-            expect(parsed[:queue]).to eq 'default'
+            expect(parsed[:max_messages]).to eq 3
             expect(parsed[:visibility_timeout]).to eq 360
           end
         end
 
         describe '#run' do
-          let(:poller) { Poller.new(['--queue', 'default', '-v', '360']) }
-
-          it 'boots rails' do
-            expect(poller).to receive(:require).with('rails')
-            expect(poller).to receive(:require).with(
-              File.expand_path('config/environment.rb')
+          let(:poller) do
+            Poller.new(
+              queue: :default,
+              visibility_timeout: 360,
+              shutdown_timeout: 42
             )
-
-            allow(poller).to receive(:poll) # no-op the poll
-            poller.run
-
-            expect(ENV.fetch('RACK_ENV', nil)).to eq 'test'
-            expect(ENV.fetch('RAILS_ENV', nil)).to eq 'test'
           end
 
           it 'merges args with loaded config' do
-            allow(poller).to receive(:boot_rails) # no-op the boot
+            expect(Aws::SQS::QueuePoller).to receive(:new).and_return(queue_poller)
 
-            allow(poller).to receive(:poll) # no-op the poll
+            expect(queue_poller).to receive(:poll).with(
+              {
+                skip_delete: true,
+                max_number_of_messages: 2, # from test app config file
+                visibility_timeout: 360 # from options
+              }
+            )
+
             poller.run
-
-            config = Aws::ActiveJob::SQS.config
-            expect(config.max_messages_for(:default)).to eq 2 # from test app config file
-            expect(config.visibility_timeout_for(:default)).to eq 360 # from argv
-            expect(config.shutdown_timeout).to eq 15 # from defaults
           end
 
           it 'polls the configured queue' do
-            allow(poller).to receive(:boot_rails) # no-op the boot
             expect(Aws::SQS::QueuePoller).to receive(:new).with(
               'https://queue-url',
               { client: instance_of(Aws::SQS::Client) }
@@ -65,7 +59,6 @@ module Aws
           end
 
           it 'runs the poller with the configured options' do
-            allow(poller).to receive(:boot_rails) # no-op the boot
             expect(Aws::SQS::QueuePoller).to receive(:new).and_return(queue_poller)
 
             expect(queue_poller).to receive(:poll).with(
@@ -80,9 +73,7 @@ module Aws
           end
 
           it 'sets max_number_of_messages to 1 for fifo queues' do
-            allow(poller).to receive(:boot_rails) # no-op the boot
-
-            allow(Aws::ActiveJob::SQS.config).to receive(:url_for).and_return('https://queue-url.fifo')
+            allow_any_instance_of(Configuration).to receive(:url_for).and_return('https://queue-url.fifo')
             expect(Aws::SQS::QueuePoller).to receive(:new).and_return(queue_poller)
 
             expect(queue_poller).to receive(:poll).with(
@@ -97,8 +88,6 @@ module Aws
           end
 
           it 'polls for messages and executes them' do
-            allow(poller).to receive(:boot_rails) # no-op the boot
-
             executor = double(Executor)
             expect(Executor).to receive(:new).and_return(executor)
 
@@ -111,15 +100,13 @@ module Aws
           end
 
           it 'calls shutdown when interrupted' do
-            allow(poller).to receive(:boot_rails) # no-op the boot
-
             executor = double(Executor)
             expect(Executor).to receive(:new).and_return(executor)
 
             expect(Aws::SQS::QueuePoller).to receive(:new).and_return(queue_poller)
             expect(queue_poller).to receive(:poll).and_raise(Poller::Interrupt)
 
-            expect(executor).to receive(:shutdown)
+            expect(executor).to receive(:shutdown).with(42) # from options
             expect(poller).to receive(:exit) # no-op the exit
 
             poller.run